+++ src/filter.rs
@@ -6,7 +6,7 @@ use log::{debug, error};
 use pest::iterators::Pair;
 use pest::Parser;
 use petgraph::algo::toposort;
-use crate::model::model_m3u::PlaylistItem;
+use crate::model::model_playlist::PlaylistItem;
 use crate::model::model_config::ItemField;
 use petgraph::graph::DiGraph;
 use crate::{create_m3u_filter_error_result};