+++ src/model/mod.rs
@@ -1,7 +1,8 @@
 pub(crate) mod config;
 pub(crate) mod model_config;
-pub(crate) mod model_m3u;
+pub(crate) mod model_playlist;
 pub(crate) mod mapping;
 pub(crate) mod api_proxy;
 pub(crate) mod stats;
 pub(crate) mod xmltv;
+pub(crate) mod model_xtream;